# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: AuraTalk Backend (Phase 1–2)
  version: 0.2.0
  description: |
    Consolidated OpenAPI spec for AuraTalk backend services.
    Phase 2 adds username identity, privacy settings, contacts, and message requests.
    NOTE: This backend stores encrypted message envelopes only (no plaintext messages).
servers:
  - url: http://localhost:8082
    description: gateway (local)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
      required: [ok, error]
...
paths:
...
  /messages:
    post:
      tags: [chat]
      summary: Send encrypted message envelope (JWT required)
      description: Server stores ciphertext only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationId:
                  type: string
                recipientUserId:
                  type: string
                  format: uuid
                recipientDeviceId:
                  type: string
                ciphertext:
                  type: string
                clientMessageId:
                  type: string
              required: [conversationId, recipientUserId, recipientDeviceId, ciphertext]
      responses:
        "200":
          description: Stored envelope id
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  envelopeId:
                    type: string
                    format: uuid
                required: [ok, envelopeId]
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /conversations/{conversationId}/messages:
    get:
      tags: [chat]
      summary: Fetch encrypted envelopes for recipient device (JWT required)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - in: query
          name: after
          required: false
          description: Return messages created after this envelope UUID (incremental sync)
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Envelopes
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        conversation_id:
                          type: string
                        sender_user_id:
                          type: string
                          format: uuid
                        sender_device_id:
                          type: string
                        ciphertext:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                        client_message_id:
                          type: string
                          nullable: true
                required: [ok, messages]
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /conversations/{conversationId}/receipts:
    post:
      tags: [chat]
      summary: Read receipt — recipient opened chat (persisted + fan-out)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: conversationId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [envelopeId]
              properties:
                envelopeId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Read receipt stored and forwarded
  /conversations:
    get:
      tags: [chat]
      summary: List 1:1 conversations for current user (JWT required)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Conversation summaries with last message ciphertext
...